+++ opennmt/data/noise.py
@@ -101,8 +101,7 @@ class WordNoiser(object):
       return tokens, sequence_length
 
 
-@six.add_metaclass(abc.ABCMeta)
-class Noise(object):
+class Noise(abc.ABC):
   """Base class for noise modules."""
 
   def __call__(self, words):