+++ opennmt/decoders/self_attention_decoder.py
@@ -23,6 +23,7 @@ class SelfAttentionDecoder(decoder.Decoder):
                ffn_activation=tf.nn.relu,
                position_encoder_class=SinusoidalPositionEncoder,
                num_sources=1,
+               maximum_relative_position=None,
                **kwargs):
     """Initializes the parameters of the decoder.
 
@@ -42,6 +43,8 @@ class SelfAttentionDecoder(decoder.Decoder):
         class to use for position encoding (or a callable that returns an
         instance).
       num_sources: The number of source contexts expected by this decoder.
+      maximum_relative_position: Maximum relative position representation
+        (from https://arxiv.org/abs/1803.02155).
       **kwargs: Additional layer arguments.
     """
     super(SelfAttentionDecoder, self).__init__(num_sources=num_sources, **kwargs)
@@ -61,7 +64,8 @@ class SelfAttentionDecoder(decoder.Decoder):
             dropout=dropout,
             attention_dropout=attention_dropout,
             ffn_dropout=ffn_dropout,
-            ffn_activation=ffn_activation)
+            ffn_activation=ffn_activation,
+            maximum_relative_position=maximum_relative_position)
         for i in range(num_layers)]
 
   @property
@@ -77,8 +81,7 @@ class SelfAttentionDecoder(decoder.Decoder):
     return True
 
   def map_v1_weights(self, weights):
-    m = []
-    m += self.output_layer.map_v1_weights(weights["dense"])
+    m = super(SelfAttentionDecoder, self).map_v1_weights(weights)
     m += self.layer_norm.map_v1_weights(weights["LayerNorm"])
     for i, layer in enumerate(self.layers):
       m += layer.map_v1_weights(weights["layer_%d" % i])