+++ opennmt/tests/scorers_test.py
@@ -1,5 +1,4 @@
 import os
-import sys
 
 import tensorflow as tf
 
@@ -11,21 +10,17 @@ class ScorersTest(tf.test.TestCase):
   def _make_perfect_hypothesis_file(self):
     ref_path = os.path.join(self.get_temp_dir(), "ref.txt")
     hyp_path = os.path.join(self.get_temp_dir(), "hyp.txt")
-    with open(ref_path, "wb") as ref_file, open(hyp_path, "wb") as hyp_file:
-      text = b"Hello world !\nHow is it going ?\n"
+    with open(ref_path, "w") as ref_file, open(hyp_path, "w") as hyp_file:
+      text = "Hello world !\nHow is it going ?\n"
       ref_file.write(text)
       hyp_file.write(text)
     return ref_path, hyp_path
 
   def testBLEUScorer(self):
-    if sys.version_info >= (3, 0):
-      bleu_scorer = scorers.BLEUScorer()
-      ref_path, hyp_path = self._make_perfect_hypothesis_file()
-      score = bleu_scorer(ref_path, hyp_path)
-      self.assertEqual(100, int(score))
-    else:
-      with self.assertRaises(ImportError):
-        bleu_scorer = scorers.BLEUScorer()
+    bleu_scorer = scorers.BLEUScorer()
+    ref_path, hyp_path = self._make_perfect_hypothesis_file()
+    score = bleu_scorer(ref_path, hyp_path)
+    self.assertEqual(100, int(score))
 
   def testROUGEScorer(self):
     rouge_scorer = scorers.ROUGEScorer()