+++ opennmt/utils/scorers.py
@@ -1,13 +1,11 @@
 """Hypotheses file scoring."""
 
 import abc
-import six
 
 import tensorflow as tf
 
 
-@six.add_metaclass(abc.ABCMeta)
-class Scorer(object):
+class Scorer(abc.ABC):
   """Scores hypotheses against references."""
 
   def __init__(self, name):
@@ -56,7 +54,7 @@ class ROUGEScorer(Scorer):
     return {"rouge-1", "rouge-2", "rouge-l"}
 
   def __call__(self, ref_path, hyp_path):
-    from rouge import FilesRouge
+    from rouge import FilesRouge  # pylint: disable=import-outside-toplevel
     files_rouge = FilesRouge(hyp_path, ref_path)
     rouge_scores = files_rouge.get_scores(avg=True)
     return {name:rouge_scores[name]["f"] for name in self.scores_name}
@@ -66,14 +64,10 @@ class BLEUScorer(Scorer):
   """Scorer using sacreBLEU."""
 
   def __init__(self):
-    try:
-      import sacrebleu  # pylint: disable=unused-import, unused-variable
-    except ImportError:
-      raise ImportError("sacreBLEU evaluator requires Python 3")
     super(BLEUScorer, self).__init__("bleu")
 
   def __call__(self, ref_path, hyp_path):
-    from sacrebleu import corpus_bleu
+    from sacrebleu import corpus_bleu  # pylint: disable=import-outside-toplevel
     with tf.io.gfile.GFile(ref_path) as ref_stream, tf.io.gfile.GFile(hyp_path) as sys_stream:
       bleu = corpus_bleu(sys_stream, [ref_stream])
       return bleu.score